% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{collectCrossValidationClass}
\alias{collectCrossValidationClass}
\title{Colects data from cross validation for gini classification.}
\usage{
collectCrossValidationClass(crossMetricsList)
}
\description{
Colects data from cross validation for gini classification.
}
