% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{createEnvironment}
\alias{createEnvironment}
\title{Creates environment to save data across functions without loss.}
\usage{
createEnvironment()
}
\value{
New Environment to save data.
}
\description{
Creates environment to save data across functions without loss.
}
\examples{
createEnvironment()
}
