% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidation}
\alias{crossValidation}
\title{Cross validation function}
\usage{
crossValidation(k, formula, attr, data, numberOfTrees, bootstrap)
}
\description{
Cross validation function
}
\examples{
crossValidation(5, quality~., 'quality', winequality.white, 30, TRUE)

}
