% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaSplit.R
\name{evalAnova}
\alias{evalAnova}
\title{Anova eval split function}
\usage{
evalAnova(y, wt, parms)
}
\arguments{
\item{y}{Target variable to predict}

\item{wt}{Tree examples weights}

\item{parms}{Optional parameters for the splitting function.}
}
\value{
Rpart init structure containing basic node information.
}
\description{
Used for printing node information by rpart.
}
