% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{evalGini}
\alias{evalGini}
\title{Gini eval split function}
\usage{
evalGini(y, wt, parms)
}
\arguments{
\item{y}{Target variable to predict}

\item{wt}{Tree examples weights}

\item{parms}{Optional parameters for the splitting function.}
}
\value{
Rpart init structure containing basic node information for printing.
}
\description{
Used for printing node information by rpart.
}
