% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{formatg}
\alias{formatg}
\title{Formatting of data for node information.}
\usage{
formatg(x, digits = getOption("digits"), format = paste0("\%.", digits, "g"))
}
\arguments{
\item{x}{Data to format}

\item{format}{Formatting of data.}

\item{digit}{Number of digits to print}
}
\value{
Formatted data.
}
\description{
Formats digists and matrixes for node information printing.
}
