% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{getClassWeights}
\alias{getClassWeights}
\title{Calculates class weights for target attribute.}
\usage{
getClassWeights(y, wt, parms)
}
\arguments{
\item{y}{Target variable to predict}

\item{wt}{Tree examples weights}

\item{parms}{Optional parameters for the splitting function.}
}
\value{
Class weights
}
\description{
Used to calculate likelihood for classes.
}
