% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{getConfusionMatrixMetricsClass}
\alias{getConfusionMatrixMetricsClass}
\title{Calculates metrics for gini classification.}
\usage{
getConfusionMatrixMetricsClass(prediction, real)
}
\description{
Calculates metrics for gini classification.
}
