% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{getGiniGoodness}
\alias{getGiniGoodness}
\title{Calculates gini split function.}
\usage{
getGiniGoodness(
  maxImpurity,
  leftClassWeights,
  leftWeights,
  rightClassWeights,
  rightWeights
)
}
\arguments{
\item{maxImpurity}{Max impurity of node.}

\item{leftClassWeights}{Class weights for left part of split}

\item{leftWeights}{Summed class weight for left part of split}

\item{rightClassWeights}{Class weights for right part of split}

\item{rightWeights}{Summed class weight for right part of split}
}
\value{
Rpart init structure containing basic split information.
}
\description{
Used for calculating split goodness.
}
