% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{giniImpurity}
\alias{giniImpurity}
\title{Calculates gini impurity}
\usage{
giniImpurity(prob)
}
\arguments{
\item{prob}{Vector of probabilites of classes.}
}
\value{
Gini impurity
}
\description{
Used for calculating split goodness.
}
