% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{graycode}
\alias{graycode}
\title{Calculates groups using graycode.}
\usage{
graycode(gray)
}
\arguments{
\item{gray}{Graycode structure containing graycode vector.}
}
\value{
Graycode structure containing next graycode number and updated graycode vector.
}
\description{
Used for splitting unique attributes.
}
