% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaSplit.R
\name{initAnova}
\alias{initAnova}
\title{Anova init split function}
\usage{
initAnova(y, offset, parms, wt)
}
\arguments{
\item{y}{Target variable to predict}

\item{offset}{Data offset}

\item{parms}{Optional parameters for the splitting function.}

\item{wt}{Tree examples weights}
}
\value{
Rpart init structure containing basic algorithm information.
}
\description{
Initializes variables for anova split function
}
