% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{initGini}
\alias{initGini}
\title{Gini init split function}
\usage{
initGini(y, offset, parms, wt)
}
\arguments{
\item{y}{Target variable to predict}

\item{offset}{Data offset}

\item{parms}{Optional parameters for the splitting function.}

\item{wt}{Tree examples weights}
}
\value{
Rpart init structure containing basic algorithm information.
}
\description{
Initializes variables for anova split function.
Code was written based on rpart source code for gini.
}
