% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{nextOrderNumber}
\alias{nextOrderNumber}
\title{Returns next number from order vector.}
\usage{
nextOrderNumber(num)
}
\arguments{
\item{num}{Number structure with order from which to take next number.}
}
\value{
Structure containing next number.
}
\description{
Used for splitting unique attributes.
}
