% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeAttributesChoice.R
\name{nodeAttributesChoice.chooseAttributes}
\alias{nodeAttributesChoice.chooseAttributes}
\title{Chooses specific attributes in tree.}
\usage{
nodeAttributesChoice.chooseAttributes(info)
}
\arguments{
\item{info}{NodeAttributesChoice structure}
}
\value{
NodeAttributesChoice structure
}
\description{
Chooses specific attributes in tree and stores its indexes to NodeAttributesChoiceInfo structure.
Adds initial attribute index from which start choosing attributes.
}
\examples{
nodeAttributesChoiceInfo <- nodeAttributesChoice.init(2, 3)
pnodeAttributesChoice.chooseAttributes(nodeAttributesChoiceInfo)

}
