% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeAttributesChoice.R
\name{nodeAttributesChoice.isAttributeChosen}
\alias{nodeAttributesChoice.isAttributeChosen}
\title{Checks if current attribute is chosen in tree.}
\usage{
nodeAttributesChoice.isAttributeChosen(info)
}
\arguments{
\item{info}{NodeAttributesChoice structure}
}
\value{
Is current attribute chosen
}
\description{
Used to iterate through tree attributes.
Increments attribute tree current index.
If current index is in chosen attribute indexes then returns True.
If not then returns False.
Function can be used to block executing split functions for not chosen attributes in tree.
}
\examples{
if (nodeAttributesChoice.isAttributeChosen(nodeAttributesChoiceInfo) {
 doIfAttributeChosen()
}

}
