% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{predict.randomForest}
\alias{predict.randomForest}
\title{Random Forest predict algorithm.}
\usage{
\method{predict}{randomForest}(
  randomForestObject,
  data,
  type = c("vector", "prob", "class"),
  na.action = na.pass
)
}
\arguments{
\item{randomForestObject}{Random forest object}

\item{data}{Data table on which to predict}
}
\value{
Random forest prediction results
}
\description{
Predicts with use of random forest algorithm.
}
\details{
Predicts on single tress of random forest and average results.
Function is associated with predict method.
}
\examples{
method <- list(eval=evalNode, split=splitNode, init=initTree)
data <- data.frame(y = numeric(10), a = numeric(10), b = numeric(10))
forest <- randomForest(y~., data, 2, numberOfTrees=3, method=method)
prediction <- predict(forest, data)

}
