% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{predictorRandomForest}
\alias{predictorRandomForest}
\title{Random forest predictor for cross validation function}
\usage{
predictorRandomForest(
  formula,
  trainSet,
  attributesToChooseCount,
  numberOfTrees,
  bootstrap,
  method,
  control
)
}
\description{
Random forest predictor for cross validation function
}
