% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{predictorRpart}
\alias{predictorRpart}
\title{Rpart prediction for cross validation function}
\usage{
predictorRpart(
  formula,
  trainSet,
  attributesToChooseCount,
  numberOfTrees,
  bootstrap,
  method,
  control
)
}
\description{
Rpart prediction for cross validation function
}
