% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{randomForest}
\alias{randomForest}
\title{Random Forest algorithm.}
\usage{
randomForest(e)
}
\arguments{
\item{e}{example}
}
\value{
None
}
\description{
Creates a random forest.
}
\examples{
randomForest()

}
