% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaSplit.R
\name{splitAnova}
\alias{splitAnova}
\title{Anova split node function}
\usage{
splitAnova(y, wt, x, parms, continuous)
}
\arguments{
\item{y}{Target variable to predict}

\item{wt}{Tree examples weights}

\item{x}{Attribute to split.}

\item{parms}{Optional parameters for the splitting function.}

\item{continuous}{Is x attribute continous flag.}
}
\value{
Rpart init structure containing basic split information.
}
\description{
Used for calculating split goodness.
}
