% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniSplit.R
\name{splitUniqueGini}
\alias{splitUniqueGini}
\title{Gini split node function for unique attribute.}
\usage{
splitUniqueGini(y, wt, x, ux, maxImpurity, parms)
}
\arguments{
\item{y}{Target variable to predict}

\item{wt}{Tree examples weights}

\item{x}{Attribute to split.}

\item{ux}{Unique x attribute on which split is performed.}

\item{maxImpurity}{Max impurity of node.}

\item{parms}{Param structure containing basic information.}
}
\value{
Rpart init structure containing basic split information.
}
\description{
Used for calculating split goodness for unique attribute.
}
