% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{summary.randomForest}
\alias{summary.randomForest}
\title{Random forest summary.}
\usage{
\method{summary}{randomForest}(randomForestObject, ...)
}
\arguments{
\item{treeObject}{Random tree object containing fitted trees structure.}
}
\value{
rpart random trees summary
}
\description{
Shows rpart summary for random forest.
}
