% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{summary.tree}
\alias{summary.tree}
\title{Tree summary.}
\usage{
\method{summary}{tree}(treeObject, ...)
}
\arguments{
\item{treeObject}{Tree object containing fitted tree structure}
}
\value{
rpart tree summary
}
\description{
Shows rpart summary for tree.
}
