% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{text.randomForest}
\alias{text.randomForest}
\title{Tree text.}
\usage{
\method{text}{randomForest}(randomForestObject, ...)
}
\arguments{
\item{treeObject}{Random tree object containing fitted trees structure.}
}
\value{
rpart random trees text
}
\description{
Shows rpart text plot for tree.
}
