% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{text.tree}
\alias{text.tree}
\title{Tree text.}
\usage{
\method{text}{tree}(treeObject, ...)
}
\arguments{
\item{treeObject}{Tree object containing fitted tree structure}
}
\value{
rpart tree text
}
\description{
Shows rpart text plot for tree.
}
