% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Creates a single random forest tree.}
\usage{
tree(
  formula,
  data,
  attributesToChooseCount,
  weights,
  subset,
  na.action = na.rpart,
  method,
  model = FALSE,
  x = FALSE,
  y = TRUE,
  parms,
  control,
  costr
)
}
\arguments{
\item{e}{example}
}
\value{
Rpart tree build using wrapped function that selects attributes.
}
\description{
Creates a single random forest tree with random selecting attributes before splitting node.
To build a tree uses rpart package.
}
\examples{
randomForest()

}
