% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{treeEvalNode}
\alias{treeEvalNode}
\title{Rpart eval wrapper function for selecting atributes in tree.}
\usage{
treeEvalNode(evalMethod, y, wt, parms)
}
\arguments{
\item{evalMethod}{User eval method - it will be executed}

\item{y}{Destination attribute to predict}

\item{wt}{Examples weights in tree}

\item{parms}{Tree parameters - contains NodeAttributesChoiceInfo structure and user additional parameters}
}
\value{
Result of user evalMethod
}
\description{
Generates indexes of attributes that has to be selected while splitting tree.
}
