% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{treeSplitNode}
\alias{treeSplitNode}
\title{Rpart split wrapper function for checking if attribute is chosen and spliting attribute.}
\usage{
treeSplitNode(splitFunction, y, wt, x, parms, continuous)
}
\arguments{
\item{y}{Destination attribute to predict}

\item{wt}{Examples weights in tree}

\item{x}{Attribute that will be splitted}

\item{parms}{Tree user parameters}

\item{continuous}{Is attribute continuous}

\item{evalMethod}{User eval method - it will be executed}
}
\value{
Result of user splitFunction or zero goodness vector of splits.
}
\description{
Checks if current index of attribute in NodeAttributesChoiceInfo structure is selected.
If attribute is chosen executes split function in argument.
Otherwise it returns no split - goodness with zero values for all splits.
}
