% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidation}
\alias{crossValidation}
\title{Cross validation prediction test function}
\usage{
crossValidation(
  k,
  formula,
  targetAttr,
  data,
  numberOfAttributes,
  numberOfTrees,
  bootstrap = TRUE,
  method = "anova",
  predictionType = "vector",
  metric = RMSE,
  collect = mean,
  changeFactor = FALSE,
  predictor = MOWRandomForest::randomForest,
  ...
)
}
\description{
Cross validation prediction test function
}
\examples{
crossValidation(5, quality~., 'quality', winequality.white, 30, TRUE)

}
