% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeAttributesChoice.R
\name{nodeAttributesChoice.init}
\alias{nodeAttributesChoice.init}
\title{Creates environment structure containing basic information about choosing attributes in trees.}
\usage{
nodeAttributesChoice.init(
  nodeAttributesChoiceInfo,
  attributesToChooseCount,
  allAttributesCount
)
}
\arguments{
\item{nodeAttributesChoiceInfo}{Environment structure to be initialized}

\item{attributesToChooseCount}{Number of attributes to choose in tree.}

\item{allAttributesCount}{Total count of attributes in tree.}
}
\value{
NodeAttributesChoiceInfo environment structure containing basic information about choosing tree attributes.
}
\description{
Creates environment structure containing basic information about choosing attributes in trees.
}
\examples{
nodeAttributesChoiceInfo <- createEnvironment()
attributesToChooseCount <- 2
allAttributesCount <- 3
nodeAttributesChoice.init(nodeAttributesChoiceInfo, attributesToChooseCount, allAttributesCount)

}
