% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Creates a single random forest tree.}
\usage{
tree(
  formula,
  data,
  attributesToChooseCount = floor(sqrt(ncol(data) - 1)),
  bootstrap = FALSE,
  na.action = na.rpart,
  method = "default",
  model = FALSE,
  parms = NULL,
  ...
)
}
\arguments{
\item{formula}{Tree formula}

\item{data}{Data table}

\item{attributesToChooseCount}{Number of attributes to choose while splitting node - default sqrt(n) where n number of attributes.}

\item{bootstrap}{Use bootstrap method - default=FALSE.}

\item{na.action}{The default action deletes all observations for which y is missing, but keeps those in which one or more predictors are missing.}

\item{method}{User split methods or specific split functions ("anova", "class", "default").}

\item{model}{Keep a copy of the model frame in the result.}

\item{parms}{Optional parameters for the splitting function.}

\item{weights}{Tree examples weights}

\item{subset}{Expression saying that only a subset of the rows of the data should be used in the fit.}

\item{x}{Keep a copy of the x matrix in the result.}

\item{y}{Keep a copy of the dependent variable in the result.}

\item{control}{A list of options that control details of the rpart algorithm.}

\item{cost}{A vector of non-negative costs.}
}
\value{
Tree structure containing rpart tree build using wrapped function that selects attributes.
}
\description{
Creates a single random forest tree with random selecting attributes before splitting node.
To build a tree uses rpart package.
}
\examples{
method <- list(eval=evalNode, split=splitNode, init=initTree)
data <- data.frame(y = numeric(10), a = numeric(10), b = numeric(10))
t <- tree(y~., data, 2, method=method)

}
